Make BranchGraph an interface.
